+++ sei-ibc-go/modules/core/05-port/keeper/keeper_test.go
@@ -44,6 +44,12 @@ func (suite *KeeperTestSuite) TestBind() {
 	capKey := suite.keeper.BindPort(suite.ctx, validPort)
 	require.NotNil(suite.T(), capKey, "capabilityKey is nil on valid BindPort")
 
+	isBound := suite.keeper.IsBound(suite.ctx, validPort)
+	require.True(suite.T(), isBound, "port is bound successfully")
+
+	isNotBound := suite.keeper.IsBound(suite.ctx, "not-a-port")
+	require.False(suite.T(), isNotBound, "port is not bound")
+
 	// Test that rebinding the same portid causes panic
 	require.Panics(suite.T(), func() { suite.keeper.BindPort(suite.ctx, validPort) }, "did not panic on re-binding the same port")
 }